Make functions parsing the event logs explicitly static.
There is no need to load a smart contract wrapper to parse an event log.

Note that this change is API (not ABI) backward compatible, I don't
think it is worth to add a separate static function, since if you regen
the wrappers is it expected that you recompile the calling code as well.